corrected PRG pin for esp32c3
sometimes, I notice things way late. Like the fact that the PRG pin on the esp32c3 is no longer GPIO0 as it was with all xtensa based chips but now it's GPIO9.
I was using GPIO9 as an output pin with a 10k pulldown which should have always set the chip into download mode - not what I wanted. I have now shuffeled pins around and this should be ok now.